+++ src/App.tsx
@@ -8,6 +8,7 @@ import { AuthProvider } from "./contexts/AuthContext";
 import { ThemeProvider } from "./contexts/ThemeContext";
 import ProtectedRoute from "./components/ProtectedRoute";
 import GlobalSparkle from "./components/GlobalSparkle";
+import OutrunGlobals from "./components/OutrunGlobals";
 import AuthPage from "./pages/AuthPage";
 import ProjectsPage from "./pages/ProjectsPage";
 import OnboardingPage from "./pages/OnboardingPage";
@@ -28,6 +29,7 @@ const App = () => (
       <ProjectProvider>
         <TooltipProvider>
           <GlobalSparkle />
+          <OutrunGlobals />
           <Toaster />
           <Sonner />
           <BrowserRouter>